requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
pydantic==2.5.0
python-dotenv==1.0.0
google-api-python-client==2.108.0
httpx==0.25.2
python-multipart==0.0.6
langchain==0.1.0
langchain-google-genai==0.0.6
streamlit==1.29.0
google-genai==0.1.0
 